+++ src/lib/connection/XMPPServiceHandler/alertEventHandler.ts
@@ -388,6 +388,9 @@ class AlertEventHandler extends GenericHandler {
                     case "joincompanyinvite":
                         // treated in invitationEventHandler
                         break;
+                    case "joincompanyrequest":
+                        // treated in invitationEventHandler
+                        break;
                     default:
                         that.logger.log("error", LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;