+++ src/lib/connection/XMPPServiceHandler/channelEventHandler.ts
@@ -160,6 +160,9 @@ class ChannelEventHandler extends GenericHandler {
                     case "joincompanyinvite":
                         // treated in invitationEventHandler
                         break;
+                    case "joincompanyrequest":
+                        // treated in invitationEventHandler
+                        break;
                     default:
                         that.logger.log("error", LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;