+++ src/lib/connection/XMPPServiceHandler/invitationEventHandler.ts
@@ -112,6 +112,9 @@ class InvitationEventHandler extends GenericHandler {
                     case "joincompanyinvite":
                         that.onJoinCompanyInviteManagementMessageReceived(node);
                         break;
+                    case "joincompanyrequest":
+                        that.onJoinCompanyRequestManagementMessageReceived(node);
+                        break;
                     default:
                         that.logger.log("error", LOG_ID + "(onManagementMessageReceived) unmanaged management message node " + node.getName());
                         break;
@@ -275,6 +278,37 @@ class InvitationEventHandler extends GenericHandler {
         } catch (err) {
             that.logger.log("error", LOG_ID + "(onJoinCompanyInviteManagementMessageReceived) CATCH Error !!! ");
             that.logger.log("internalerror", LOG_ID + "(onJoinCompanyInviteManagementMessageReceived) CATCH Error !!! : ", err);
+            return true;
+        }
+    }
+
+    onJoinCompanyRequestManagementMessageReceived (stanza) {
+        let that = this;
+        that.logger.log("internal", LOG_ID + "(onJoinCompanyRequestManagementMessageReceived) stanza : ", "\n", stanza.root ? prettydata.xml(stanza.root().toString()) : stanza);
+
+        try {
+            let joincompanyrequestElem = stanza; //.find("userinvite");
+            if (joincompanyrequestElem && joincompanyrequestElem.attrs) {
+                let id = joincompanyrequestElem.attrs.id;
+                let type = joincompanyrequestElem.attrs.type;
+                let action = joincompanyrequestElem.attrs.action;
+                let status = joincompanyrequestElem.attrs.status;
+
+                let request = {
+                    id,
+                    type,
+                    action,
+                    status
+                };
+                that.eventEmitter.emit("evt_internal_joinCompanyRequestManagementUpdate", request);
+                return true;
+            } else {
+                that.logger.log("error", LOG_ID + "(onJoinCompanyRequestManagementMessageReceived) joincompanyinvite empty.");
+                that.logger.log("internalerror", LOG_ID + "(onJoinCompanyRequestManagementMessageReceived) joincompanyinvite empty : ", stanza);
+            }
+        } catch (err) {
+            that.logger.log("error", LOG_ID + "(onJoinCompanyRequestManagementMessageReceived) CATCH Error !!! ");
+            that.logger.log("internalerror", LOG_ID + "(onJoinCompanyRequestManagementMessageReceived) CATCH Error !!! : ", err);
             return true;
         }
     }